requirements.txt
asgiref==3.8.1
astroid==3.3.10
dill==0.4.0
Django==4.2.21
isort==6.0.1
mccabe==0.7.0
platformdirs==4.3.8
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-decouple==3.8
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
djangorestframework==3.16.0
requests==2.32.3
drf-yasg==1.21.10
